refactor: reorganize subscription event types and improve type imports
